move user and user_config serializers and tests
