+++ src/sentry/api/endpoints/auth_login.py
@@ -6,8 +6,8 @@ from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import Endpoint, control_silo_endpoint
 from sentry.api.serializers.base import serialize
-from sentry.api.serializers.models.user import DetailedSelfUserSerializer
 from sentry.models.organization import Organization
+from sentry.users.api.serializers.user import DetailedSelfUserSerializer
 from sentry.utils import auth, metrics
 from sentry.utils.hashlib import md5_text
 from sentry.web.forms.accounts import AuthenticationForm
