+++ src/sentry/api/endpoints/team_members.py
@@ -10,13 +10,13 @@ from sentry.api.bases.team import TeamEndpoint
 from sentry.api.paginator import OffsetPaginator
 from sentry.api.serializers import Serializer, register, serialize
 from sentry.api.serializers.models.organization_member.response import OrganizationMemberResponse
-from sentry.api.serializers.models.user import UserSerializerResponse
 from sentry.apidocs.constants import RESPONSE_FORBIDDEN, RESPONSE_NOT_FOUND
 from sentry.apidocs.examples.team_examples import TeamExamples
 from sentry.apidocs.parameters import CursorQueryParam, GlobalParams
 from sentry.apidocs.utils import inline_sentry_response_serializer
 from sentry.models.organizationmember import InviteStatus
 from sentry.models.organizationmemberteam import OrganizationMemberTeam
+from sentry.users.api.serializers.user import UserSerializerResponse
 
 
 class OrganizationMemberOnTeamResponse(OrganizationMemberResponse):
