+++ src/sentry/api/serializers/models/__init__.py
@@ -67,8 +67,6 @@ from .sentry_app_installation import *  # noqa: F401,F403
 from .servicehook import *  # noqa: F401,F403
 from .tagvalue import *  # noqa: F401,F403
 from .team import *  # noqa: F401,F403
-from .user import *  # noqa: F401,F403
-from .user_identity_config import *  # noqa: F401,F403
 from .user_social_auth import *  # noqa: F401,F403
 from .useremail import *  # noqa: F401,F403
 from .userip import *  # noqa: F401,F403
