+++ src/sentry/api/serializers/models/discoversavedquery.py
@@ -2,9 +2,9 @@ from collections import defaultdict
 from typing import DefaultDict, TypedDict
 
 from sentry.api.serializers import Serializer, register
-from sentry.api.serializers.models.user import UserSerializerResponse
 from sentry.constants import ALL_ACCESS_PROJECTS
 from sentry.discover.models import DatasetSourcesTypes, DiscoverSavedQuery, DiscoverSavedQueryTypes
+from sentry.users.api.serializers.user import UserSerializerResponse
 from sentry.users.services.user.service import user_service
 from sentry.utils.dates import outside_retention_with_modified_start, parse_timestamp
