+++ src/sentry/api/serializers/models/organization_member/response.py
@@ -7,8 +7,8 @@ from sentry.api.serializers.models.role import (
     OrganizationRoleSerializerResponse,
     TeamRoleSerializerResponse,
 )
-from sentry.api.serializers.models.user import UserSerializerResponse
 from sentry.integrations.api.serializers.models.external_actor import ExternalActorResponse
+from sentry.users.api.serializers.user import UserSerializerResponse
 
 
 class SCIMName(TypedDict):
