+++ src/sentry/users/api/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .user import *  # noqa: F401,F403
+from .user_identity_config import *  # noqa: F401,F403
