+++ tests/sentry/users/api/serializers/test_user_identity_config.py
@@ -2,11 +2,11 @@ from datetime import timedelta
 from unittest import mock
 
 from sentry.api.serializers import serialize
-from sentry.api.serializers.models.user_identity_config import Status, UserIdentityConfig
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.authprovider import AuthProvider
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
+from sentry.users.api.serializers.user_identity_config import Status, UserIdentityConfig
 from sentry.users.models.identity import Identity
 from social_auth.models import UserSocialAuth
